feat: add infinite scroll hook
